Move QSS template reading tests to test_theme_utils
read_qss_stylesheet_file only names a resource and delegates to
read_theming_resource, so its tests belong with the other resource
loader tests rather than among the template processing ones.
